Finished adding authentication to endpoints
