+++ training/training/routes/tabular/tabular.py
@@ -8,11 +8,15 @@ from torch.utils.data import DataLoader
 from training.core.optimizer import getOptimizer
 from training.core.trainer import ClassificationTrainer, RegressionTrainer
 from training.routes.tabular.schemas import TabularParams
+from core.authenticator import FirebaseAuth
 
 router = Router()
 
 
-@router.post("")
+@router.post(
+    "",
+    auth=FirebaseAuth()
+)
 def tabularTrain(request: HttpRequest, tabularParams: TabularParams):
     if tabularParams.default:
         dataCreator = SklearnDatasetCreator.fromDefault(
